chore: 迁移前快照（VRC OSC 模块 + UI 迭代 + openvr 链接库）
- 新增 crates/vox-osc-win（VRChat OSC 控制）
- 新增 docs/VRC_OSC_PROTOCOL.md
- UI: 字幕/管线卡片/i18n/样式多处迭代
- core: cloud/pipeline/runtime/settings 调整
- gitignore: 排除 target-vr-check/（本地 2.3GB 构建产物）

+++ crates/vox-audio-win/src/capture/mic.rs
@@ -17,7 +17,7 @@ pub(crate) fn open_microphone(device_name: Option<&str>) -> PortResult<OpenCaptu
     open_capture_client(
         eCapture,
         device_name,
-        AUDCLNT_STREAMFLAGS_EVENTCALLBACK as u32,
+        AUDCLNT_STREAMFLAGS_EVENTCALLBACK,
         "麦克风",
     )
 }
